Reorganize network code
