+++ ethereum/p2p/src/main/java/tech/pegasys/pantheon/ethereum/p2p/network/netty/TimeoutHandler.java
@@ -10,7 +10,7 @@
  * an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
  * specific language governing permissions and limitations under the License.
  */
-package tech.pegasys.pantheon.ethereum.p2p.netty;
+package tech.pegasys.pantheon.ethereum.p2p.network.netty;
 
 import java.util.concurrent.TimeUnit;
 import java.util.function.Supplier;
@@ -51,7 +51,7 @@ public class TimeoutHandler<C extends Channel> extends ChannelInitializer<C> {
   }
 
   @FunctionalInterface
-  interface OnTimeoutCallback {
+  public interface OnTimeoutCallback {
     void invoke();
   }
 }
